feat: list habit logs endpoint
logs now carry their own unit and daily target so old entries stay accurate after a habit is edited

listing returns a habit's full history newest first, scoped to the owner

count gets clamped to the target on update instead of being rejected

+++ migrations/000003_create_habits.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS habits (
     id           UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
     user_id      UUID        NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     name         TEXT        NOT NULL,
-    exp_reward   BIGINT      NOT NULL,
-    daily_target BIGINT      NOT NULL,
+    exp_reward   INTEGER     NOT NULL,
+    daily_target INTEGER     NOT NULL,
     unit         TEXT        NOT NULL,
     created_at   TIMESTAMPTZ NOT NULL DEFAULT now(),
     deleted_at   TIMESTAMPTZ
